Putting and reachable.
I changed the location in the code where we have the algo account for the extra amount rolled. I ended up using 8% extra roll arbitrarily but also I didn't want to account for the whole 10% when looking at the splash zone. I also added +1 to the putting distance. This definitely isn't consistent in getting the putter to make it in one shot.